fix(integration): Move TransferType: `objective` -> `argument`
