+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/thebrewingproject/argument/TransferType.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.compatibility.thebrewingproject.objective;
+package org.betonquest.betonquest.compatibility.thebrewingproject.argument;
 
 /**
  * Define a brew transfer type between a player and a structure.
