+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/thebrewingproject/objective/BrewTransferObjectiveFactory.java
@@ -15,6 +15,7 @@ import org.betonquest.betonquest.api.quest.objective.ObjectiveFactory;
 import org.betonquest.betonquest.api.quest.objective.service.ObjectiveService;
 import org.betonquest.betonquest.compatibility.thebrewingproject.argument.BrewQualityArgument;
 import org.betonquest.betonquest.compatibility.thebrewingproject.argument.BrewingStructureType;
+import org.betonquest.betonquest.compatibility.thebrewingproject.argument.TransferType;
 
 /**
  * Factory for {@link BrewTransferObjective}.
